Added button with convert .xls file with msk coordinates to kml file. Classes AdvTable and Dialog transported to separate files
